requirements/dev.txt
-r base.txt

pip-tools
django-debug-toolbar
django-debug-toolbar-template-profiler